+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [2026-05-16 22:36] data/data/com.termux/files/home/godmode/lib/encoder.py
 - [2026-05-16 22:37] home/godmode/lib/parsers.py
 - [2026-05-16 22:38] home/godmode/engine/racer.py
+- [2026-05-16 22:39] data/data/com.termux/files/home/godmode/engine/scorer.py